Rename Lib* components: drop Lib prefix, Btn → Button
LibBtn → Button, LibCard → Card, LibField → Field,
LibNotice → Notice, LibStamp → Stamp. Update all imports
and template references accordingly.
