+++ web/src/components/Notice.vue
