+++ web/src/components/Stamp.vue
